Complete delete product feature for admin user

+++ products/views.py
@@ -151,9 +151,12 @@ def edit_product(request, product_id):
 def delete_product(request, product_id):
     """ Deletes a product from the store """
     product = get_object_or_404(Product, pk=product_id)
-    product.delete()
-    messages.success(request, 'Product deleted successfully!')
-    return redirect(reverse('products'))
+    if request.method == 'POST':
+        product.delete()
+        messages.success(request, 'Product deleted successfully!')
+        return redirect('products')
+        
+    return render(request, 'products/delete_product.html')
 
 
 # Add Comment
